+++ include/netAbstract.h
@@ -11,7 +11,7 @@ namespace net
         virtual net_data get_net_data() = 0;
         virtual std::vector<DATA_TYPE> launch_forward(const std::vector<DATA_TYPE> &inputs) = 0; //* returns result
         virtual void init_gradient(const net_sets &sets) = 0;                                    
-        virtual std::vector<DATA_TYPE> launch_gradient(int iterations) = 0;                      //* returns iterations errors
+        virtual std::vector<DATA_TYPE> launch_gradient(size_t iterations) = 0;                      //* returns iterations errors
         virtual void print_inner_vals() = 0;
         virtual signed long get_gradient_performance() = 0;
         virtual signed long get_forward_performance() = 0;
